Moved metric related files into the metric folder
